ml_service/requirements.txt
# --- Core ---
fastapi==0.115.0
uvicorn[standard]==0.30.1

# --- Config ---
pydantic==2.7.4
pydantic-settings==2.3.3
python-dotenv==1.0.1

# --- HTTP ---
httpx==0.27.0

# --- DB ---
sqlalchemy==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9

# --- Migrations ---
alembic==1.13.1

# --- Auth / Security ---
python-jose==3.3.0
passlib[bcrypt]==1.7.4

# --- ML ---
numpy==1.26.4
scikit-learn==1.5.0
pandas==2.2.2
joblib==1.4.2

# --- Utils / Logs ---
loguru==0.7.2
